part-4  Color Image Compression/differential_lzw/requirements.txt
pillow>=9.0.0
numpy>=1.20.0 